moved make_halls methods from GameMapBSP to GameMap since they can apply to all GameMaps changed GameMap.make_room() to use generic Space object
